Add database functionality and /debit and /credit functions.
The script is now able to use a database to store the expenditure data. The code that interacts with the database  is in the
SQL_utils.py file.
The function /expenditure was replaced with /debit and /credit which let you decide whether you are adding or substracting money
from a particular account.
There are a lot of try/except clauses now to handle exceptions.