Add requirements.txt
Using the command pip freeze > requirments.txt I added all the modules I was using in python to a .txt file. Then I checked which of them
were being used by the bot and deleted the ones that weren't. Finally I created a virtual enviroment and install the modules that remained
in the requirements.txt.