Implement messages after /debit and /credit commands.
Error and success messages were added to the main.py code so that the user knows what happend with their requests. In order to implent this
it was necessary to change the code a bit. Now the try/except clauses are in the add_debit and add_credit definitions and not in
add_expenditure.  A new exception name was created and the try/except clauses were changed  bit.