New command: /create_account
Added a new command that allows the user to add a new account to the database. I had to create a new SQL command to do that and a
unique constraint to the database so that there's no repeated account names. In the future we could change this by storing the telegram
user id. The new code also handles exceptions with try/except clauses and I created a new exception class in the SQL_utils.py file.